deviq-graphrag/pyproject.toml
[project]
name = "deviq-graphrag"
version = "0.1.0"
description = "GraphRAG-powered skill discovery for DevIQ"
requires-python = ">=3.13"
dependencies = [
    "lightrag-hku>=1.0.0",
    "qdrant-client>=1.12",
    "fastapi>=0.115",
    "uvicorn[standard]",
    "openai>=1.60",
    "python-frontmatter>=1.1",
    "pyyaml>=6.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
